Wrap long lines
It looks like we changed a lot of code here, but that isn't the case.
The only change required to annatar to allow word wrap was that we now
use artano v0.2.0 instead of v0.1.0. The other changes are all of a
similar variety; cargo update and so forth.

I also ran cargo fmt.